Sketch out OWSIdentityManager.
// FREEBIE
