+++ src/Messages/OWSIdentityManager.h
@@ -3,15 +3,33 @@
 //
 
 #import <AxolotlKit/IdentityKeyStore.h>
-#import "TSStorageManager.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@class OWSRecipientIdentity;
-
 extern NSString *const TSStorageManagerTrustedKeysCollection;
 
-@interface TSStorageManager (IdentityKeyStore) <IdentityKeyStore>
+// extern NSString *const kNSNotificationName_VerificationStateDidChange;
+
+@class OWSRecipientIdentity;
+
+typedef NS_ENUM(NSUInteger, OWSVerificationState) {
+    OWSVerificationStateDefault,
+    OWSVerificationStateVerified,
+    OWSVerificationStateNoLongerVerified,
+};
+
+// This class can be safely accessed and used from any thread.
+@interface OWSIdentityManager : NSObject <IdentityKeyStore>
+
+- (instancetype)init NS_UNAVAILABLE;
+
++ (instancetype)sharedManager;
+
+//- (void)setVerificationState:(OWSVerificationState)verificationState
+//              forPhoneNumber:(NSString *)phoneNumber
+//       isUserInitiatedChange:(BOOL)isUserInitiatedChange;
+//
+//- (OWSVerificationState)verificationStateForPhoneNumber:(NSString *)phoneNumber;
 
 /**
  * Explicitly mark an identity as approved for blocking/nonblocking use
@@ -26,10 +44,10 @@ extern NSString *const TSStorageManagerTrustedKeysCollection;
  * @returns YES if we are replacing an existing known identity key for recipientId.
  *          NO  if there was no previously stored identity key for the recipient.
  */
-- (BOOL)saveRemoteIdentity:(NSData *)identityKey
-               recipientId:(NSString *)recipientId
-    approvedForBlockingUse:(BOOL)approvedForBlockingUse
- approvedForNonBlockingUse:(BOOL)approvedForNonBlockingUse;
+//- (BOOL)saveRemoteIdentity:(NSData *)identityKey
+//               recipientId:(NSString *)recipientId
+//    approvedForBlockingUse:(BOOL)approvedForBlockingUse
+// approvedForNonBlockingUse:(BOOL)approvedForNonBlockingUse;
 
 /**
  * Check if a recipient identity corresponds to an untrusted identity
@@ -38,7 +56,7 @@ extern NSString *const TSStorageManagerTrustedKeysCollection;
  * @returns nil if the identity doesn't exist or if it's trusted
  *          else returns the untrusted identity
  */
-- (nullable OWSRecipientIdentity *)unconfirmedIdentityThatShouldBlockSendingForRecipientId:(NSString *)recipientId;
+//- (nullable OWSRecipientIdentity *)unconfirmedIdentityThatShouldBlockSendingForRecipientId:(NSString *)recipientId;
 
 /**
  * @param   recipientId unique stable identifier for the recipient, e.g. e164 phone number
@@ -46,11 +64,17 @@ extern NSString *const TSStorageManagerTrustedKeysCollection;
  *          else returns the unseen identity
  *
  */
-- (nullable OWSRecipientIdentity *)unseenIdentityChangeForRecipientId:(NSString *)recipientId;
+//- (nullable OWSRecipientIdentity *)unseenIdentityChangeForRecipientId:(NSString *)recipientId;
+
+/**
+ * @param   recipientId unique stable identifier for the recipient, e.g. e164 phone number
+ * @returns if the recipient's current identity is trusted.
+ */
+- (BOOL)isCurrentIdentityTrustedForSendingWithRecipientId:(NSString *)recipientId;
 
 - (void)generateNewIdentityKey;
 - (nullable NSData *)identityKeyForRecipientId:(NSString *)recipientId;
-- (void)removeIdentityKeyForRecipient:(NSString *)receipientId;
+//- (void)removeIdentityKeyForRecipient:(NSString *)receipientId;
 
 @end
